[TASK] Initial commit
Meta data of an EC2 instance can be retrieved by accessing an
AWS-internal URI (169.254.169.254). The instance type (for example
"t2.large") is one of the meta data available.

This TYPO3 extension retrieves the instance type from the meta data and
shows the result in TYPO3's system information toolbar in the backend.